Manipulate cmdline history in s:processcmdline().
Add the un-expanded, user-typed, command line to the history and attempt
to delete the expanded version afterwards. CmdlineLeave is triggered
before leaving command line, and deleting a history entry at this point
would delete the carefully recorded un-expanded version. ShellCmdPost is
used as a workaround. This creates room for bugs if the user types
<marked>, <yanked>, etc. not in shell commands. However, room for such
bugs is already plentiful thanks to s:resetmarkedonsuccess.