Optimize constructing lists of paths from lines.
Constructing the path of a file based on its line number is performed by
going up the whole directory tree and visiting multiple, if not all,
lines above. When several lines are to be converted into paths,
filemanager may visit the same lines multiple times. To avoid this,
a separate function is created. This function is not replacing the
normal s:undercursor() since the new function is considered heavier on
the memory.

Also remove unused variables from s:undercursor().

Also rename a local variable in s:undercursor() for clarity.

Also use v:numbermax to initialize depth in s:undercursor().