Slightly change regex matching.
The only change is that regexes in :Filter, :Mark, and :Yank are now
matched against relative paths with '/' prepended to them. This
simplifies matching names starting with a specific regex.

Before, to match names starting with 'abc' in the tree root directory or
deeper, one had to use '\(^\|/\)abc'. Now '/abc' is enough. '^abc' would
be converted to '/abc'. To match files only in the tree root directory,
use '^/abc'.

Changes to the code:

* Prepend '/' to matched strings in s:filtercontents()
  and s:namematches().

* Create a separate function for regex conversion.
  New: convert '^' to '/'; don't convert '^/'

* Use s:convertpattern() in s:filtercmd(), s:markbypat(),
  s:checkconfig().

* Convert all keys of g:filemanager_sortrules in s:checkconfig().