Fix bug with glob() usage.
Vim's glob() function uses 'wildignorecase' and 'fileignorecase'.

When 'fileignorecase' is set, 'wildignorecase' is ignored.
However, on my system this results in case never being ignored, which is
very counter-intuitive. So on my system, the only way to ignore case is
to set 'wildignorecase' and 'nofileignorecase'.

Moreover, when there is nothing to expand (the glob expression contains
no wildcards), glob() seems to not use 'wildignorecase', either.

Given all these peculiarities, it is only fitting to always set
'nowildignorecase' and 'nofileignorecase' before calls to glob().

Hence the following changes are made.

* Create a separate function to check if a path exists.

* Use this function where a single item is checked: s:dirreadable(),
  s:writebookmarks(), s:bookmarkdel(), s:newdir(), s:open().

* Reset and restore ignore case options in functions where glob()
  is called several times: s:getdircontents(), s:pastemarked(),
  s:renamebylist().

* The use of this function in s:printcontents() will be changed soon.
  The decision to not reset ignore case options is due to the fact that
  the user may interrupt printing the directory tree (if for some reason
  it takes too long), in which case these options would not be restored
  correctly.

* Also don't return lists from glob() where strings are enough.