Change history deletion in s:processcmdline().
Add the auto command to delete the latest entry on ShellCmdPost only
when the command line actually contains an "!".

This limitation ensures that a shell command will definitely run, and
the only history entry at the risk is the one generated by filemanager.

The un-expanded, user-typed command line is always saved, just like
before.